+++ composeApp/src/iosMain/kotlin/eu/heha/samayouwa/MainViewController.kt
@@ -2,4 +2,4 @@ package eu.heha.samayouwa
 
 import androidx.compose.ui.window.ComposeUIViewController
 
-fun MainViewController() = ComposeUIViewController { App() }
+fun MainViewController() = ComposeUIViewController { App.Content() }