+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [1.0.1] - 2021-XX-XX
+## [1.1.0] - 2021-12-09
 
 - pressing space now toggles play/pause
 - pressing keys now triggers the quick settings
@@ -9,12 +9,15 @@
 - improved animation of quick settings
 - renamed libs and plugins
 - extracted variables
+- added desktop icons
+- added persistence of settings
+- overhauled build process for desktop apps
 
 ## [1.0.0] - 2024-11-27
 
 - basic interpolation of a ball going from right to left and back in a linear fashion
 - added play/pause functionality
-- added settings for
+- added set~~~~tings for
     - velocity
     - size of the ball
     - color of the ball